pyproject.toml
[tool.isort]
py_version = 39
profile = 'black'

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py39']

[tool.poetry]
name = "fastbrownieapi"
version = "0.1.0"
description = ""
authors = ["oregu <mefouiny@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.5"
fastapi = "^0.110.0"
ipfshttpclient = "^0.7.0"
pillow = "^10.2.0"
eth-brownie = "^1.20.2"
openai = "^1.14.3"
uvicorn = "0.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
